refactor(Home):move the Home's Main component to separeted directory to improve components organization and refresh all imports tree
