+++ src/pages/Home.jsx
@@ -1,6 +1,6 @@
 import { Divider } from "@chakra-ui/react";
 import Header from "../components/Header";
-import Main from "../components/Main";
+import Main from "../components/Home/Main";
 import Footer from "../components/Footer";
